Fix window.open windowName
Hi,
null as windowName works, but not in all use cases. when e.g. sharing via Google+ and you are not logged into your google account, opening the popup will windowName = null will open the popup, login to google+ but then the window closes (as there is no name to return to), so you can login but cannot share...
Setting the windowName to "" fixes this faulty behavior, you can now login to google in your popup and after loggin in, the popup stays open and allows you to share the page :)